Add TLS and authentication support
Previously, the connection to the MQTT broker was completely unencrypted.
This commit fixes that by adding TLS support together with username and
password authentication for access control. Required is now a TLS
certificate root CA in PEM format included in the configuration header,
and the current time is fetched from a global NTP server in order to
check the certificate validity.

Also improves the interpretation of the messages on the interval topic.